Make `abort` private and add a note about posssible unsafety.
Calling `abort` can result in an unsafe state of the output port
because `close` should be called immediately after aborting. The only
place where `close` is called, is when the output port is dropped.
Therefore the port had to be reinstantiated to open it again.
The benefit of providing an `abort` function is not high enough to
compensate the downsides.