Refactor tests
You cant rely on the default Midi port to be available in a test runs.
It's not a test failure if an error occurs, therefore ouput the result
values and go on. To see the test outputs run it like this:
`cargo test -- --no-capture`
The test should only fail on segfaults, which it does if one occurs.